# apps/grc-platform/backend/openapi.yaml
openapi: 3.0.3
info:
  title: GRC Platform Backend API
  description: Public-facing API for the GRC Platform (Audit Hub and Risk Hub). Authenticates
    callers via a Bearer JWT issued by Asgardeo and validated against the issuer's JWKS
    (RS256, iss/aud checked). Most data access is proxied to the internal Compliance
    Entity service; file uploads are proxied through this backend to Azure Blob Storage
    rather than issuing direct SAS URLs to clients.
  version: 1.0.0
servers:
- url: http://localhost:8081
  description: Local development
...
security:
- bearerAuth: []
paths:
...
  /api/v1/risks:
    get:
      tags:
      - Risks
      summary: List risks
      description: Spreadsheet-style column filtering — every multi-value field is OR-matched within itself and AND-matched against the others. A caller who holds only CompleteActionSteps (Action Owner) is scoped to risks where they own an action plan.
      operationId: listRisks
      parameters:
      - name: statuses
        in: query
        description: Comma-separated workflow_status values.
        schema:
          type: string
      - name: team_id
        in: query
        description: Comma-separated source_register_id values.
        schema:
          type: string
      - name: level
        in: query
        description: Comma-separated LOW | MEDIUM | HIGH values.
        schema:
          type: string
      - name: search
        in: query
        description: Matched against risk_code and risk_title.
        schema:
          type: string
      - name: risk_type
        in: query
        description: Comma-separated NEW | UPDATED values.
        schema:
          type: string
      - name: owner_id
        in: query
        description: Comma-separated owner user IDs.
        schema:
          type: string
      - name: submitted_from
        in: query
        description: created_at >= this date (YYYY-MM-DD).
        schema:
          type: string
          format: date
      - name: submitted_to
        in: query
        description: created_at <= this date (YYYY-MM-DD).
        schema:
          type: string
          format: date
      - name: due_from
        in: query
        description: implementation_date >= this date (YYYY-MM-DD).
        schema:
          type: string
          format: date
      - name: due_to
        in: query
        description: implementation_date <= this date (YYYY-MM-DD).
        schema:
          type: string
          format: date
      - name: due_overdue
        in: query
        description: '"true" to additionally restrict to implementation_date < today.'
        schema:
          type: string
          enum:
          - 'true'
          - 'false'
      - name: open_escalation
        in: query
        description: '"true" restricts to risks carrying an unresolved (OPEN) escalation — what the Overdue Risks tab filters on.'
        schema:
          type: string
          enum:
          - 'true'
          - 'false'
      - name: limit
        in: query
        schema:
          type: integer
          default: 50
          maximum: 200
      - name: offset
        in: query
        schema:
          type: integer
          default: 0
      responses:
        '200':
          description: Paginated risk list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskListPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
      - Risks
      summary: Create a risk
      description: Atomically generates a risk code and creates the risk, action plan, steps, compliance-reference links, and change-log entry inside one DB transaction.
      operationId: createRisk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRiskRequest'
      responses:
        '201':
          description: Risk created
          headers:
            Location:
              description: /api/v1/risks/{id}
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRiskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/risks/{id}:
    get:
      tags:
      - Risks
      summary: Get a risk by ID
      operationId: getRisk
      parameters:
      - $ref: '#/components/parameters/RiskId'
      responses:
        '200':
          description: Risk detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
      - Risks
      summary: Update a risk
      description: Changing implementation_date, email_subject, or action_steps on an IN_REMEDIATION risk moves it to PENDING_AMENDMENT. All other fields are free-edit.
      operationId: updateRisk
      parameters:
      - $ref: '#/components/parameters/RiskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRiskRequest'
      responses:
        '204':
          description: Risk updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/risks/{id}/owner-approve:
    post:
      tags:
      - Risk Workflow
      summary: Risk owner's approval
      description: Handles PENDING_RISK_OWNER_APPROVAL, PENDING_AMENDMENT, and PENDING_OWNER_COMPLETION_APPROVAL.
      operationId: ownerApproveRisk
      parameters:
      - $ref: '#/components/parameters/RiskId'
      responses:
        '204':
          description: Approved
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/risks/{id}/management-approve:
    post:
      tags:
      - Risk Workflow
      summary: Management approval
      description: PENDING_MANAGEMENT_APPROVAL → PENDING_COMPLIANCE_REVIEW.
      operationId: managementApproveRisk
      parameters:
      - $ref: '#/components/parameters/RiskId'
      responses:
        '204':
          description: Approved
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Asgardeo-issued JWT, validated by signature (RS256 via JWKS), issuer, and audience. Sent as `Authorization Bearer <token>`. Not required for GET /health.
  parameters:
...
    RiskId:
      name: id
      in: path
      required: true
      description: Risk ID.
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The caller lacks the privilege required for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The resource is not in a state that allows this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: The request is well-formed but refers to something that could not be verified (e.g. an unrecognised employee email)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: Uploaded file exceeds the 25 MiB limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      description: Standard JSON error envelope returned on every non-2xx response.
      properties:
        message:
          type: string
...
    RiskScore:
      type: object
      description: One cell of the 3×3 likelihood × impact matrix.
      properties:
        id:
          type: integer
        likelihood:
          type: integer
        impact:
          type: integer
        risk_rating:
          type: integer
        risk_level:
          type: string
        color_code:
          type: string
    ComplianceReference:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
...
    CreateActionStepRequest:
      type: object
      properties:
        description:
          type: string
    CreateRiskRequest:
      type: object
      required:
      - year
      - quarter
      - source_register_id
      - risk_title
      - risk_description
      - identified_by_type
      - assigner_id
      - risk_identified_date
      - likelihood
      - impact
      - impact_description
      - implementation_date
      - reassessment_date
      - assignment_team_id
      - owner_id
      - management_approver_id
      - action_owner_id
      - action_plan_description
      - action_steps
      - treatment_strategy
      - email_subject
      properties:
        year:
          type: integer
        quarter:
          type: string
          enum:
          - Q1
          - Q2
          - Q3
          - Q4
        source_register_id:
          type: integer
        risk_title:
          type: string
        risk_description:
          type: string
        compliance_reference_ids:
          type: array
          items:
            type: integer
        identified_by_type:
          type: string
          enum:
          - EMPLOYEE
          - EXTERNAL_PERSON
          - TOOL
        identified_by_name:
          type: string
          nullable: true
          description: Ignored when identified_by_type is EMPLOYEE — the server derives it from identified_by_email instead.
        identified_by_email:
          type: string
          nullable: true
          description: Required when identified_by_type is EMPLOYEE.
        assigner_id:
          type: integer
        risk_identified_date:
          type: string
          format: date
        risk_category_ids:
          type: array
          items:
            type: integer
          description: Writes risk_category_reference rows. Genuinely many-to-many, though the Add Risk form only ever sends one today via a single-select dropdown.
        likelihood:
          type: integer
          minimum: 1
          maximum: 3
        impact:
          type: integer
          minimum: 1
          maximum: 3
        impact_description:
          type: string
        implementation_date:
          type: string
          format: date
        reassessment_date:
          type: string
          format: date
        assignment_team_id:
          type: integer
        owner_id:
          type: integer
        management_approver_id:
          type: integer
          description: Required on every risk regardless of level or treatment strategy — names who approves PENDING_MANAGEMENT_APPROVAL/PENDING_MANAGEMENT_CLOSURE_APPROVAL (ACCEPT+HIGH risks only) and who an ESCALATED HIGH risk's escalation is answered by.
        action_owner_id:
          type: integer
        action_plan_description:
          type: string
        action_steps:
          type: array
          items:
            $ref: '#/components/schemas/CreateActionStepRequest'
        treatment_strategy:
          type: string
        progress:
          type: string
        git_issue_url:
          type: string
        email_subject:
          type: string
        remarks:
          type: string
    CreateRiskResponse:
      type: object
      properties:
        id:
          type: integer
        risk_code:
          type: string
    RiskListItem:
      type: object
      description: Lightweight row DTO for GET /api/v1/risks; joins resolve display names so the table needs no secondary fetches.
      properties:
        id:
          type: integer
        risk_code:
          type: string
        risk_title:
          type: string
        source_register_name:
          type: string
        risk_level:
          type: string
        risk_level_color:
          type: string
        owner_name:
          type: string
        assigner_name:
          type: string
        workflow_status:
          type: string
        risk_type:
          type: string
          enum:
          - NEW
          - UPDATED
        implementation_date:
          type: string
          format: date
          nullable: true
        rejection_comment:
          type: string
          nullable: true
        rejection_stage:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    RiskListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RiskListItem'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    ActionPlanStep:
      type: object
      properties:
        id:
          type: integer
        plan_id:
          type: integer
        step_no:
          type: integer
        description:
          type: string
          nullable: true
        status:
          type: string
        completed_date:
          type: string
          format: date
          nullable: true
    ActionPlanDetail:
      type: object
      description: ActionPlan with its steps embedded, used inside RiskDetail.
      properties:
        id:
          type: integer
        action_owner_id:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
        plan_type:
          type: string
          enum:
          - STANDARD
          - MANAGEMENT
        steps:
          type: array
          items:
            $ref: '#/components/schemas/ActionPlanStep'
    RiskAssessment:
      type: object
      description: One residual risk reassessment entry.
      properties:
        id:
          type: integer
        risk_id:
          type: integer
        score_id:
          type: integer
        progress:
          type: string
        reassessment_date:
          type: string
          format: date
        assessed_by:
          type: string
        created_at:
          type: string
          format: date-time
        residual_likelihood:
          type: integer
        residual_impact:
          type: integer
        residual_rating:
          type: integer
        residual_level:
          type: string
        residual_color_code:
          type: string
        is_initial:
          type: boolean
          description: Marks a synthetic entry representing the risk's gross score, prepended so the assessment log shows the full lineage.
    RiskDetail:
      type: object
      description: Enriched DTO returned by GET /api/v1/risks/{id}.
      properties:
        id:
          type: integer
        risk_code:
          type: string
        risk_year:
          type: integer
        risk_quarter:
          type: string
        risk_title:
          type: string
        risk_description:
          type: string
        risk_identified_date:
          type: string
          format: date
          nullable: true
        identified_by_type:
          type: string
          nullable: true
        identified_by_name:
          type: string
          nullable: true
        assigner_id:
          type: integer
        owner_id:
          type: integer
        management_approver_id:
          type: integer
        impact_description:
          type: string
          nullable: true
        treatment_strategy:
          type: string
          nullable: true
        source_register_id:
          type: integer
        assignment_team_id:
          type: integer
        progress:
          type: string
          nullable: true
        implementation_date:
          type: string
          format: date
          nullable: true
        reassessment_date:
          type: string
          format: date
          nullable: true
        git_issue_url:
          type: string
          nullable: true
        email_subject:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        workflow_status:
          type: string
        risk_type:
          type: string
          enum:
          - NEW
          - UPDATED
        rejection_comment:
          type: string
          nullable: true
        rejection_stage:
          type: string
          nullable: true
        owner_first_approved_at:
          type: string
          format: date-time
          nullable: true
        compliance_approval_date:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        source_register_name:
          type: string
        assignment_team_name:
          type: string
        owner_name:
          type: string
        assigner_name:
          type: string
        management_approver_name:
          type: string
        compliance_approver_name:
          type: string
          nullable: true
        gross_score:
          nullable: true
          anyOf:
          - $ref: '#/components/schemas/RiskScore'
          description: The original rating assigned at creation; immutable once a risk owner has approved the risk.
        effective_score:
          nullable: true
          anyOf:
          - $ref: '#/components/schemas/RiskScore'
          description: The current residual score — the latest reassessment's score when one exists, else the gross score. This is what tables/headers should display.
        compliance_references:
          type: array
          items:
            $ref: '#/components/schemas/ComplianceReference'
        risk_categories:
          type: array
          items:
            $ref: '#/components/schemas/RiskCategory'
        action_plan:
          nullable: true
          anyOf:
          - $ref: '#/components/schemas/ActionPlanDetail'
        assessments:
          type: array
          items:
            $ref: '#/components/schemas/RiskAssessment'
    UpdateActionStepRequest:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        description:
          type: string
    UpdateRiskRequest:
      type: object
      required:
      - risk_title
      - risk_description
      - email_subject
      properties:
        risk_title:
          type: string
        risk_description:
          type: string
        risk_identified_date:
          type: string
          format: date
        identified_by_type:
          type: string
          description: Empty/omitted means "leave Identified By unchanged".
          enum:
          - ''
          - EMPLOYEE
          - EXTERNAL_PERSON
          - TOOL
        identified_by_name:
          type: string
          nullable: true
        identified_by_email:
          type: string
          nullable: true
        assigner_id:
          type: integer
          nullable: true
        owner_id:
          type: integer
          nullable: true
        management_approver_id:
          type: integer
          nullable: true
        impact_description:
          type: string
        compliance_reference_ids:
          type: array
          items:
            type: integer
        risk_category_ids:
          type: array
          items:
            type: integer
        progress:
          type: string
        git_issue_url:
          type: string
        remarks:
          type: string
        treatment_strategy:
          type: string
        assignment_team_id:
          type: integer
          nullable: true
        action_plan_description:
          type: string
        action_owner_id:
          type: integer
          nullable: true
        implementation_date:
          type: string
          format: date
          description: Restricted — changing this on an IN_REMEDIATION risk triggers PENDING_AMENDMENT.
        email_subject:
          type: string
          description: Restricted — changing this on an IN_REMEDIATION risk triggers PENDING_AMENDMENT.
        action_steps:
          type: array
          description: Restricted — changing this on an IN_REMEDIATION risk triggers PENDING_AMENDMENT.
          items:
            $ref: '#/components/schemas/UpdateActionStepRequest'
        reassessment_date:
          type: string
          format: date
          description: Full-edit only (editable before risk owner approval).
        gross_score_id:
          type: integer
          nullable: true
          description: Full-edit only (editable before risk owner approval).
...
    RiskCategory:
      type: object
      description: A fixed-but-extensible risk classification (e.g. "PII / Sensitive Data Exposure").
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
...